- did code reorganization; - added weak reference to self in closure; - func updateProgress() reconfigured with .setProgress(...); - JSONDecoder  reconfigured with .convertFromSnakeCase; - reconfigured  progressView.progressViewStyle  from .bar to .default
