+++ ImageFeed/Protocols+Models/AuthViewControllerDelegateProtocol.swift
@@ -3,3 +3,4 @@ import UIKit
 protocol AuthViewControllerDelegateProtocol: AnyObject {
     func didAuthenticate(_ vc: AuthViewController)
 }
+
