+++ ImageFeed/Protocols+Models/OAuthTokenResponseBodyModel.swift
@@ -5,15 +5,7 @@ struct OAuthTokenResponseBody: Codable {
     let tokenType: String
     let scope: String
     let createdAt: Int
-    
-    enum CodingKeys: String, CodingKey {
-        case accessToken = "access_token"
-        case tokenType = "token_type"
-        case scope = "scope"
-        case createdAt = "created_at"
-    }
 }
-
 
 /*https://unsplash.com/documentation/user-authentication-workflow
  If successful, the response body will be a JSON representation of your user’s access token:
