+++ ImageFeed/SplashViewController/SplashViewController.swift
@@ -2,28 +2,30 @@ import UIKit
 
 class SplashViewController: UIViewController {
     
+    //MARK: - Private Properties
+    
     private let showAuthenticationScreenSegueIdentifier = "showAuthView"
     private let showImageListScreenSegueIdentifier = "showImageListView"
     
+    //MARK: - Public Methods
+    
     override func viewDidAppear(_ animated: Bool) {
-        
         super.viewDidAppear(animated)
         //отладочное
         //UserDefaults.standard.setValue("test", forKey: OAuth2TokenStorage.shared.tokenKey)
         //UserDefaults.standard.removeObject(forKey: OAuth2TokenStorage.shared.tokenKey)
         
         completedAuthorizationCheck(accessToken: OAuth2TokenStorage.shared.token)
-        
     }
     
+    //MARK: - Private Methods
+    
     private func completedAuthorizationCheck(accessToken: String?) {
-        
         if let _ = accessToken {
             switchToTabBarController()
         } else {
             performSegue(withIdentifier: showAuthenticationScreenSegueIdentifier, sender: self)
         }
-        
     }
     
     private func switchToTabBarController() {
@@ -39,15 +41,16 @@ class SplashViewController: UIViewController {
         
         // Установим в `rootViewController` полученный контроллер
         window.rootViewController = tabBarController
+        
     }
     
 }
 
+//MARK: - Prepare for segue
+
 extension SplashViewController {
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        
         if segue.identifier == showAuthenticationScreenSegueIdentifier {
-            
             guard
                 let navigationController = segue.destination as? UINavigationController,
                 let viewController = navigationController.viewControllers[0] as? AuthViewController
@@ -55,17 +58,16 @@ extension SplashViewController {
                 assertionFailure("SplashViewController (33) - Failed to prepare for \(showAuthenticationScreenSegueIdentifier)")
                 return
             }
-            
             viewController.delegate = self
-            
         } else {
             super.prepare(for: segue, sender: sender)
         }
     }
 }
 
+//MARK: - AuthViewControllerDelegateProtocol
+
 extension SplashViewController: AuthViewControllerDelegateProtocol {
-    
     func didAuthenticate(_ vc: AuthViewController) {
         switchToTabBarController()
         vc.dismiss(animated: true)
